Revises metadata input test
